Added deletion confirm

+++ server/app/users/routes.py
@@ -1,13 +1,15 @@
-import os
-
-from flask import current_app, flash, redirect, render_template, request, url_for
+from flask import flash, redirect, render_template, request, url_for
 from flask_login import current_user, login_required, login_user, logout_user
 from itsdangerous import BadData, SignatureExpired
 from werkzeug.security import check_password_hash, generate_password_hash
 
 from ..extensions import db, serializer
-from ..funcs import delete_user_files, send_verification_email
-from ..models import User
+from ..funcs import (
+    delete_files_from_bucket,
+    get_filenames_from_attachments,
+    send_verification_email,
+)
+from ..models import CardSet, FlashCard, User
 from . import bp
 from .forms import LoginForm, RegisterForm
 
@@ -19,7 +21,10 @@ def login():
 
     form = LoginForm()
     if form.validate_on_submit():
-        user = User.query.filter_by(username=form.username.data).first()
+        user = (
+            User.query.filter_by(username=form.username.data).first()
+            or User.query.filter_by(email=form.username.data).first()
+        )
         if user:
             if check_password_hash(user.password, form.password.data):
                 login_user(user, remember=True)
@@ -57,7 +62,9 @@ def register():
         token = serializer.dumps(user_data, salt="email-confirm").encode("utf-8")
         verification_url = url_for(".confirm_email", token=token, _external=True)
 
-        send_verification_email(verification_url, email)
+        send_verification_email(
+            email, "mails/verification.html", "Verify your email", url=verification_url
+        )
         return render_template(
             "common/error.html",
             error="Verification link has been send to your email successfully",
@@ -106,12 +113,27 @@ def update_email(token):
     return redirect(url_for("main.main"))
 
 
-@bp.route("/delete-profile")
+@bp.route("/delete-account/<token>")
 @login_required
-def delete_profile():
-    delete_user_files(current_user.id)
-    db.session.delete(current_user)
-    db.session.commit()
-    logout_user()
-    flash("Your account was deleted successfully", "success")
-    return redirect(url_for("users.register"))
+def delete_account(token):
+    try:
+        user_info = serializer.loads(token, salt="delete-account", max_age=3600)
+    except SignatureExpired:
+        return render_template("common/error.html", error="Verification token expired!")
+    except BadData:
+        return render_template("common/error.html", error="Invalid verification token")
+    if current_user.id == user_info["id"]:
+        user_attachments = (
+            db.session.query(FlashCard.attachments)
+            .join(CardSet)
+            .filter(CardSet.user_id == current_user.id)
+            .all()
+        )
+        filenames = get_filenames_from_attachments(user_attachments)
+        filenames and delete_files_from_bucket(filenames)
+        db.session.delete(current_user)
+        db.session.commit()
+        logout_user()
+        flash("Your account was deleted successfully", "success")
+        return redirect(url_for("users.register"))
+    return redirect(url_for("users.login"))
